school_env/school_env/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "openenv-school-timetable"
version = "1.0.0"
description = "School Timetable Scheduling — OpenEnv Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "pydantic>=2.0,<3.0",
    "fastapi>=0.104",
    "uvicorn>=0.24",
]

[project.scripts]
server = "server.app:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "server*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.md"]
